chore: shift indexes for examples
