+++ projects/demo/src/modules/components/select/examples/10/index.html
@@ -1,0 +1,13 @@
+<tui-textfield tuiChevron>
+    <input
+        placeholder="Select something"
+        tuiSelect
+        [formControl]="control"
+    />
+
+    <tui-data-list-wrapper
+        *tuiTextfieldDropdown
+        new
+        [items]="items"
+    />
+</tui-textfield>
