+++ projects/demo/src/modules/components/select/examples/5/index.html
@@ -1,26 +1,26 @@
-<tui-textfield tuiChevron>
+<tui-textfield
+    tuiChevron
+    [content]="content"
+>
     <input
-        placeholder="Country"
         tuiSelect
-        [(ngModel)]="value"
+        [formControl]="control"
     />
 
-    <cdk-virtual-scroll-viewport
-        *tuiTextfieldDropdown
-        tuiScrollable
-        class="scroll"
-        [itemSize]="44"
-    >
-        <tui-data-list>
-            <button
-                *cdkVirtualFor="let item of countries | async"
-                new
-                tuiOption
-                type="button"
-                [value]="item"
-            >
-                {{ item }}
-            </button>
-        </tui-data-list>
-    </cdk-virtual-scroll-viewport>
+    <tui-data-list *tuiTextfieldDropdown>
+        <button
+            *ngFor="let item of items"
+            new
+            tuiOption
+            type="button"
+            [value]="item.id"
+        >
+            {{ item.name }}
+        </button>
+    </tui-data-list>
 </tui-textfield>
+
+<p>
+    <strong>Form control:</strong>
+    <code>{{ control.value | json }}</code>
+</p>
