+++ projects/demo/src/modules/components/select/examples/6/index.ts
@@ -1,31 +1,37 @@
-import {Component} from '@angular/core';
+import {
+    CdkFixedSizeVirtualScroll,
+    CdkVirtualForOf,
+    CdkVirtualScrollViewport,
+} from '@angular/cdk/scrolling';
+import {AsyncPipe} from '@angular/common';
+import {Component, inject} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
-import {TuiDropdownMobile} from '@taiga-ui/addon-mobile';
-import type {TuiBooleanHandler} from '@taiga-ui/cdk';
-import {TuiPlatform} from '@taiga-ui/cdk';
-import {TuiTextfield} from '@taiga-ui/core';
-import {TuiChevron, TuiDataListWrapper, TuiSelect} from '@taiga-ui/kit';
+import {TuiDataList, TuiScrollable, TuiTextfield} from '@taiga-ui/core';
+import {TUI_COUNTRIES, TuiChevron, TuiSelect} from '@taiga-ui/kit';
+import {map} from 'rxjs';
 
 @Component({
     standalone: true,
     imports: [
+        AsyncPipe,
+        CdkFixedSizeVirtualScroll,
+        CdkVirtualForOf,
+        CdkVirtualScrollViewport,
         FormsModule,
         TuiChevron,
-        TuiDataListWrapper,
-        TuiDropdownMobile,
-        TuiPlatform,
+        TuiDataList,
+        TuiScrollable,
         TuiSelect,
         TuiTextfield,
     ],
     templateUrl: './index.html',
+    styleUrls: ['./index.less'],
     encapsulation,
     changeDetection,
 })
 export default class Example {
-    protected readonly platforms = ['web', 'ios', 'android'] as const;
-    protected value: 'android' | 'ios' | 'web' | null = 'ios';
-    protected readonly disabledItemHandler: TuiBooleanHandler<string> = (x) =>
-        x === 'web';
+    protected readonly countries = inject(TUI_COUNTRIES).pipe(map(Object.values));
+    protected value = null;
 }
