+++ projects/demo/src/modules/components/select/examples/7/index.ts
@@ -1,25 +1,21 @@
-import {NgIf} from '@angular/common';
-import {Component, inject} from '@angular/core';
-import {FormControl, ReactiveFormsModule} from '@angular/forms';
+import {Component} from '@angular/core';
+import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
-import type {TuiBooleanHandler, TuiStringHandler} from '@taiga-ui/cdk';
-import {TUI_IS_MOBILE} from '@taiga-ui/cdk';
+import {TuiDropdownMobile} from '@taiga-ui/addon-mobile';
+import type {TuiBooleanHandler} from '@taiga-ui/cdk';
+import {TuiPlatform} from '@taiga-ui/cdk';
 import {TuiTextfield} from '@taiga-ui/core';
 import {TuiChevron, TuiDataListWrapper, TuiSelect} from '@taiga-ui/kit';
-
-interface Character {
-    id: number;
-    name: string;
-}
 
 @Component({
     standalone: true,
     imports: [
-        NgIf,
-        ReactiveFormsModule,
+        FormsModule,
         TuiChevron,
         TuiDataListWrapper,
+        TuiDropdownMobile,
+        TuiPlatform,
         TuiSelect,
         TuiTextfield,
     ],
@@ -28,18 +24,8 @@ interface Character {
     changeDetection,
 })
 export default class Example {
-    protected readonly isMobile = inject(TUI_IS_MOBILE);
-    protected readonly control = new FormControl(null);
-    protected items: readonly Character[] = [
-        {name: 'Luke Skywalker', id: 1},
-        {name: 'Leia Organa Solo', id: 2},
-        {name: 'Darth Vader', id: 3},
-        {name: 'Han Solo', id: 4},
-        {name: 'Obi-Wan Kenobi', id: 5},
-        {name: 'Yoda', id: 6},
-    ];
-
-    protected stringify: TuiStringHandler<Character> = (item) => item.name;
-    protected readonly disabledItemHandler: TuiBooleanHandler<Character> = (item) =>
-        item.name === 'Darth Vader';
+    protected readonly platforms = ['web', 'ios', 'android'] as const;
+    protected value: 'android' | 'ios' | 'web' | null = 'ios';
+    protected readonly disabledItemHandler: TuiBooleanHandler<string> = (x) =>
+        x === 'web';
 }
