+++ projects/demo/src/modules/components/select/examples/8/index.ts
@@ -1,23 +1,45 @@
-import {Component} from '@angular/core';
+import {NgIf} from '@angular/common';
+import {Component, inject} from '@angular/core';
 import {FormControl, ReactiveFormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
+import type {TuiBooleanHandler, TuiStringHandler} from '@taiga-ui/cdk';
+import {TUI_IS_MOBILE} from '@taiga-ui/cdk';
 import {TuiTextfield} from '@taiga-ui/core';
-import {TuiChevron, TuiSelect} from '@taiga-ui/kit';
+import {TuiChevron, TuiDataListWrapper, TuiSelect} from '@taiga-ui/kit';
+
+interface Character {
+    id: number;
+    name: string;
+}
 
 @Component({
     standalone: true,
-    imports: [ReactiveFormsModule, TuiChevron, TuiSelect, TuiTextfield],
+    imports: [
+        NgIf,
+        ReactiveFormsModule,
+        TuiChevron,
+        TuiDataListWrapper,
+        TuiSelect,
+        TuiTextfield,
+    ],
     templateUrl: './index.html',
     encapsulation,
     changeDetection,
 })
 export default class Example {
-    protected groupItems: ReadonlyArray<readonly string[]> = [
-        ['Caesar', 'Greek', 'Apple and Chicken'],
-        ['Broccoli Cheddar', 'Chicken and Rice', 'Chicken Noodle'],
+    protected readonly isMobile = inject(TUI_IS_MOBILE);
+    protected readonly control = new FormControl(null);
+    protected items: readonly Character[] = [
+        {name: 'Luke Skywalker', id: 1},
+        {name: 'Leia Organa Solo', id: 2},
+        {name: 'Darth Vader', id: 3},
+        {name: 'Han Solo', id: 4},
+        {name: 'Obi-Wan Kenobi', id: 5},
+        {name: 'Yoda', id: 6},
     ];
 
-    protected readonly labels = ['Salad', 'Soup'];
-    protected readonly control = new FormControl<string | null>(null);
+    protected stringify: TuiStringHandler<Character> = (item) => item.name;
+    protected readonly disabledItemHandler: TuiBooleanHandler<Character> = (item) =>
+        item.name === 'Darth Vader';
 }
