--- projects/demo/src/modules/components/select/examples/99/index.html
@@ -1,26 +1,0 @@
-<tui-textfield
-    tuiChevron
-    [content]="content"
->
-    <input
-        tuiSelect
-        [formControl]="control"
-    />
-
-    <tui-data-list *tuiTextfieldDropdown>
-        <button
-            *ngFor="let item of items"
-            new
-            tuiOption
-            type="button"
-            [value]="item.id"
-        >
-            {{ item.name }}
-        </button>
-    </tui-data-list>
-</tui-textfield>
-
-<p>
-    <strong>Form control:</strong>
-    <code>{{ control.value | json }}</code>
-</p>
