+++ projects/demo/src/modules/components/select/index.html
@@ -147,8 +147,8 @@
         <tui-doc-example
             id="form-control-output"
             heading="Choose form control output"
-            [component]="99 | tuiComponent"
-            [content]="99 | tuiExample"
+            [component]="5 | tuiComponent"
+            [content]="5 | tuiExample"
             [description]="todoDescription"
         >
             <ng-template #todoDescription>
@@ -172,8 +172,8 @@
         <tui-doc-example
             id="virtual-scroll"
             heading="Virtual scroll"
-            [component]="5 | tuiComponent"
-            [content]="5 | tuiExample"
+            [component]="6 | tuiComponent"
+            [content]="6 | tuiExample"
             [description]="virtualScrollDescription"
         >
             <ng-template #virtualScrollDescription>
@@ -194,8 +194,8 @@
         <tui-doc-example
             id="dropdown-mobile"
             heading="With DropdownMobile"
-            [component]="6 | tuiComponent"
-            [content]="6 | tuiExample"
+            [component]="7 | tuiComponent"
+            [content]="7 | tuiExample"
             [description]="dropdownMobileDescription"
         >
             <ng-template #dropdownMobileDescription>
@@ -234,8 +234,8 @@
         <tui-doc-example
             id="native-picker-disabled-option"
             heading="Native picker with disabled option"
-            [component]="7 | tuiComponent"
-            [content]="7 | tuiExample"
+            [component]="8 | tuiComponent"
+            [content]="8 | tuiExample"
             [description]="nativePickerDescription"
         >
             <ng-template #nativePickerDescription>
@@ -252,15 +252,15 @@
         <tui-doc-example
             id="native-picker-groups"
             heading="Native picker with grouping options"
-            [component]="8 | tuiComponent"
-            [content]="8 | tuiExample"
+            [component]="9 | tuiComponent"
+            [content]="9 | tuiExample"
         />
 
         <tui-doc-example
             id="override-option-component"
             heading="Override option component"
-            [component]="9 | tuiComponent"
-            [content]="9 | tuiExample: 'html,ts,less' : {'option.ts': selectOptionExample}"
+            [component]="10 | tuiComponent"
+            [content]="10 | tuiExample: 'html,ts,less' : {'option.ts': selectOptionExample}"
             [description]="overrideOptionDescription"
         >
             <ng-template #overrideOptionDescription>
